refactor(runtime): re-home endpoint wiring package
Move the endpoint and TMCI wiring cluster into network.crypta.runtime.endpoints and update affected imports, package statements, and tests.\n\nCarry the existing SpotBugs exclusions forward so the package move remains tooling-neutral.
