+++ src/main/java/network/crypta/node/ClientContextInitParams.java
@@ -8,6 +8,8 @@ import network.crypta.client.async.DatastoreChecker;
 import network.crypta.client.async.USKManager;
 import network.crypta.crypt.MasterSecret;
 import network.crypta.crypt.RandomSource;
+import network.crypta.runtime.endpoints.NodeClientCoreInit;
+import network.crypta.runtime.endpoints.NodeClientPersistence;
 import network.crypta.support.MemoryLimitedJobRunner;
 import network.crypta.support.PriorityAwareExecutor;
 import network.crypta.support.Ticker;
