+++ src/main/java/network/crypta/node/Node.java
@@ -41,6 +41,7 @@ import network.crypta.node.subsystem.NodeNetworkSubsystem;
 import network.crypta.node.subsystem.NodeRoutingSubsystem;
 import network.crypta.node.subsystem.NodeServicesSubsystem;
 import network.crypta.node.subsystem.NodeStorageSubsystem;
+import network.crypta.runtime.endpoints.NodeClientCoreInit;
 import network.crypta.support.Fields;
 import network.crypta.support.HexUtil;
 import network.crypta.support.PriorityAwareExecutor;
