+++ src/main/java/network/crypta/node/NodeClientCore.java
@@ -23,6 +23,10 @@ import network.crypta.keys.NodeSSK;
 import network.crypta.node.SecurityLevels.PHYSICAL_THREAT_LEVEL;
 import network.crypta.node.runtime.LegacyRuntimePorts;
 import network.crypta.node.useralerts.UserAlertManager;
+import network.crypta.runtime.endpoints.ClientEndpoints;
+import network.crypta.runtime.endpoints.NodeClientCoreInit;
+import network.crypta.runtime.endpoints.NodeClientPersistence;
+import network.crypta.runtime.endpoints.TextModeClientInterface;
 import network.crypta.runtime.spi.RuntimePorts;
 import network.crypta.support.Base64;
 import network.crypta.support.MemoryLimitedJobRunner;
