+++ src/main/java/network/crypta/node/NodeClientCoreSupport.java
@@ -40,6 +40,7 @@ import network.crypta.node.useralerts.DiskSpaceUserAlert;
 import network.crypta.node.useralerts.SimpleUserAlert;
 import network.crypta.node.useralerts.UserAlert;
 import network.crypta.node.useralerts.UserAlertManager;
+import network.crypta.runtime.endpoints.NodeClientCoreInit;
 import network.crypta.support.SizeUtil;
 import network.crypta.support.compress.Compressor;
 import network.crypta.support.io.FileUtil;
