+++ src/main/java/network/crypta/runtime/admin/LegacyQueueInsertPort.java
@@ -28,10 +28,10 @@ import network.crypta.clients.fcp.FcpInsertTuningOptions;
 import network.crypta.clients.fcp.IdentifierCollisionException;
 import network.crypta.clients.fcp.NotAllowedException;
 import network.crypta.keys.FreenetURI;
-import network.crypta.node.ClientEndpoints;
 import network.crypta.node.Node;
 import network.crypta.node.NodeClientCore;
 import network.crypta.node.RequestStarter;
+import network.crypta.runtime.endpoints.ClientEndpoints;
 import network.crypta.runtime.spi.QueueBrowserUploadInsertRequest;
 import network.crypta.runtime.spi.QueueInsertFailureReason;
 import network.crypta.runtime.spi.QueueInsertOutcome;
