+++ src/main/java/network/crypta/runtime/endpoints/NodeClientPersistence.java
@@ -1,4 +1,4 @@
-package network.crypta.node;
+package network.crypta.runtime.endpoints;
 
 import java.io.File;
 import java.util.Objects;
@@ -15,6 +15,13 @@ import network.crypta.clients.fcp.PersistentRequestRoot;
 import network.crypta.config.Option;
 import network.crypta.config.SubConfig;
 import network.crypta.crypt.MasterSecret;
+import network.crypta.node.ClientContextInitParams;
+import network.crypta.node.ConfigurablePersister;
+import network.crypta.node.ConfigurablePersisterParams;
+import network.crypta.node.Node;
+import network.crypta.node.NodeClientCore;
+import network.crypta.node.NodeInitException;
+import network.crypta.node.Persistable;
 import network.crypta.runtime.spi.RuntimePorts;
 import network.crypta.support.SimpleFieldSet;
 import network.crypta.support.Ticker;
