+++ src/main/java/network/crypta/runtime/endpoints/TextModeClientInterfaceServer.java
@@ -1,4 +1,4 @@
-package network.crypta.node;
+package network.crypta.runtime.endpoints;
 
 import java.io.File;
 import java.io.IOException;
@@ -18,6 +18,9 @@ import network.crypta.config.SubConfig;
 import network.crypta.crypt.RandomSource;
 import network.crypta.io.NetworkInterface;
 import network.crypta.io.SSLNetworkInterface;
+import network.crypta.node.Node;
+import network.crypta.node.NodeClientCore;
+import network.crypta.node.RequestStarter;
 import network.crypta.node.runtime.SSL;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
