+++ src/test/java/network/crypta/clients/http/CoreHttpShellRuntimeSupportTest.java
@@ -10,7 +10,6 @@ import network.crypta.client.async.ClientContext;
 import network.crypta.clients.http.bookmark.BookmarkManager;
 import network.crypta.clients.http.bookmark.CoreBookmarkRuntimeSupport;
 import network.crypta.config.PersistentConfig;
-import network.crypta.node.ClientEndpoints;
 import network.crypta.node.Node;
 import network.crypta.node.NodeClientCore;
 import network.crypta.node.RequestStarter;
@@ -21,6 +20,7 @@ import network.crypta.node.SecurityLevels;
 import network.crypta.node.subsystem.NodeNetworkSubsystem;
 import network.crypta.node.subsystem.NodeServicesSubsystem;
 import network.crypta.node.useralerts.UserAlertManager;
+import network.crypta.runtime.endpoints.ClientEndpoints;
 import network.crypta.runtime.spi.RuntimePorts;
 import network.crypta.support.Ticker;
 import org.junit.jupiter.api.Test;
