+++ src/test/java/network/crypta/clients/http/SimpleToadletServerTest.java
@@ -15,10 +15,10 @@ import network.crypta.config.PersistentConfig;
 import network.crypta.config.SubConfig;
 import network.crypta.io.NetworkInterface;
 import network.crypta.io.SSLNetworkInterface;
-import network.crypta.node.ClientEndpoints;
 import network.crypta.node.NodeClientCore;
 import network.crypta.node.RequestStarter;
 import network.crypta.node.useralerts.UserAlertManager;
+import network.crypta.runtime.endpoints.ClientEndpoints;
 import network.crypta.runtime.spi.RandomnessPort;
 import network.crypta.runtime.spi.RuntimePorts;
 import network.crypta.support.HTMLNode;
