+++ src/test/java/network/crypta/node/NodeClientCoreTest.java
@@ -22,6 +22,8 @@ import network.crypta.node.SecurityLevels.PHYSICAL_THREAT_LEVEL;
 import network.crypta.node.runtime.LegacyRuntimePorts;
 import network.crypta.node.subsystem.NodeServicesSubsystem;
 import network.crypta.node.useralerts.UserAlertManager;
+import network.crypta.runtime.endpoints.ClientEndpoints;
+import network.crypta.runtime.endpoints.NodeClientPersistence;
 import network.crypta.runtime.spi.RuntimePorts;
 import network.crypta.support.PriorityAwareExecutor;
 import network.crypta.support.SimpleFieldSet;
