+++ src/test/java/network/crypta/runtime/admin/LegacyQueueDownloadPortTest.java
@@ -5,8 +5,8 @@ import network.crypta.client.async.PersistenceDisabledException;
 import network.crypta.clients.fcp.FCPServer;
 import network.crypta.clients.fcp.NotAllowedException;
 import network.crypta.clients.fcp.PersistentGlobalRequestParams;
-import network.crypta.node.ClientEndpoints;
 import network.crypta.node.NodeClientCore;
+import network.crypta.runtime.endpoints.ClientEndpoints;
 import network.crypta.runtime.spi.QueueDownloadRejectedException;
 import network.crypta.runtime.spi.QueueDownloadRequest;
 import network.crypta.runtime.spi.RequestQueueUnavailableException;
