+++ src/test/java/network/crypta/runtime/admin/LegacyQueueInsertPortTest.java
@@ -39,8 +39,8 @@ import network.crypta.config.Config;
 import network.crypta.crypt.MasterSecret;
 import network.crypta.crypt.RandomSource;
 import network.crypta.node.ClientContextResources;
-import network.crypta.node.ClientEndpoints;
 import network.crypta.node.NodeClientCore;
+import network.crypta.runtime.endpoints.ClientEndpoints;
 import network.crypta.runtime.spi.QueueBrowserUploadInsertRequest;
 import network.crypta.runtime.spi.QueueInsertFailureReason;
 import network.crypta.runtime.spi.QueueInsertOptions;
