+++ src/test/java/network/crypta/runtime/admin/LegacyQueueMutationPortTest.java
@@ -6,8 +6,8 @@ import network.crypta.clients.fcp.DownloadRequestStatus;
 import network.crypta.clients.fcp.FCPServer;
 import network.crypta.clients.fcp.RequestStatus;
 import network.crypta.clients.fcp.UploadFileRequestStatus;
-import network.crypta.node.ClientEndpoints;
 import network.crypta.node.NodeClientCore;
+import network.crypta.runtime.endpoints.ClientEndpoints;
 import network.crypta.runtime.spi.RequestQueueUnavailableException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
