+++ src/test/java/network/crypta/runtime/admin/LegacyQueuePagePortTest.java
@@ -11,11 +11,11 @@ import network.crypta.clients.fcp.RequestStatus;
 import network.crypta.clients.fcp.UploadFileRequestStatus;
 import network.crypta.clients.fcp.UploadRequestStatus;
 import network.crypta.keys.FreenetURI;
-import network.crypta.node.ClientEndpoints;
 import network.crypta.node.DarknetPeerNode;
 import network.crypta.node.Node;
 import network.crypta.node.NodeClientCore;
 import network.crypta.node.RequestStarterGroup;
+import network.crypta.runtime.endpoints.ClientEndpoints;
 import network.crypta.runtime.spi.QueuePageRequest;
 import network.crypta.runtime.spi.QueuePageSnapshot;
 import network.crypta.runtime.spi.RequestQueueUnavailableException;
