+++ src/test/java/network/crypta/runtime/admin/LegacyQueueSupportPortTest.java
@@ -3,10 +3,10 @@ package network.crypta.runtime.admin;
 import java.io.File;
 import java.nio.file.Path;
 import network.crypta.clients.fcp.FCPServer;
-import network.crypta.node.ClientEndpoints;
 import network.crypta.node.Node;
 import network.crypta.node.NodeClientCore;
 import network.crypta.node.subsystem.NodeStorageSubsystem;
+import network.crypta.runtime.endpoints.ClientEndpoints;
 import network.crypta.runtime.spi.QueuePersistenceStatusSnapshot;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
