+++ src/test/java/network/crypta/runtime/endpoints/ClientEndpointsTest.java
@@ -1,10 +1,13 @@
-package network.crypta.node;
+package network.crypta.runtime.endpoints;
 
 import network.crypta.client.async.ClientContext;
 import network.crypta.clients.fcp.FCPServer;
 import network.crypta.clients.http.FProxyToadlet;
 import network.crypta.clients.http.SimpleToadletServer;
 import network.crypta.config.Config;
+import network.crypta.node.Node;
+import network.crypta.node.NodeClientCore;
+import network.crypta.node.NodeClientCoreSupport;
 import network.crypta.node.useralerts.UserAlert;
 import network.crypta.node.useralerts.UserAlertManager;
 import network.crypta.runtime.spi.RuntimePorts;
