+++ src/test/java/network/crypta/runtime/endpoints/NodeClientCoreInitTest.java
@@ -1,4 +1,4 @@
-package network.crypta.node;
+package network.crypta.runtime.endpoints;
 
 import network.crypta.clients.http.SimpleToadletServer;
 import network.crypta.config.Config;
